No more iife and some restructuring.
