+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules
 dist
+build
 **/*.js
 !examples/**/*.js
 !rollup.config.js
