+++ examples/nanocal.html
@@ -4,14 +4,13 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <link rel="stylesheet" href="../dist/nanocal.min.css">
-  <link rel="stylesheet" href="../common/src/style.css">
+  <link rel="stylesheet" href="../build/nanocal.min.css">
   <title>Nanocal</title>
 </head>
 <body>
   <div id="nanocal"></div>
   <script type="module">
-    import Nanocal from '../dist/nanocal.js'
+    import Nanocal from '../build/nanocal.js'
     const nanocal = new Nanocal({ target: document.querySelector('#nanocal') })
     nanocal.on('selectedDay', (day) => console.log(day))
   </script>
