+++ nanocal/src/DatePicker.html
@@ -13,9 +13,9 @@
 
 <script>
   import { calendarMonth } from 'cntdys'
-  import { dayClass, weekClass } from '../../common/src/index.js'
-  import Wrapper from '../../common/src/Wrapper.html'
-  import Header from '../../common/src/Header.html'
+  import { dayClass, weekClass } from '../../common/index.js'
+  import Wrapper from '../../common/Wrapper.html'
+  import Header from '../../common/Header.html'
 
   export default {
     helpers: {
