+++ ranger/package.json
@@ -1,9 +1,7 @@
 {
   "name": "nanocal-ranger",
   "version": "0.2.8",
-  "main": "dist/ranger.cjs.js",
   "module": "dist/ranger.min.esm.js",
-  "unpkg": "dist/ranger.min.iife.js",
   "repository": "https://github.com/zigomir/nanocal/tree/master/ranger",
   "license": "MIT",
   "homepage": "https://github.com/zigomir/nanocal",
@@ -11,18 +9,15 @@
   "description": "Minimal, AirBnb inspired, range picker",
   "scripts": {
     "prebuild": "rm -f dist/*",
-    "build": "../node_modules/.bin/tsc && npm run build:cjs && npm run build:esm && npm run build:esm:min && npm run build:iife && npm run build:iife:min && ../node_modules/.bin/bundlesize",
-    "build:cjs": "../node_modules/.bin/rollup src/main.js -c ../rollup.config.js -f cjs -o dist/ranger.cjs.js",
+    "build": "../node_modules/.bin/tsc && npm run build:esm && npm run build:esm:min && ../node_modules/.bin/bundlesize",
     "build:esm": "../node_modules/.bin/rollup src/main.js -c ../rollup.config.js -f es -o dist/ranger.esm.js",
     "build:esm:min": "../node_modules/.bin/rollup src/main.js -c ../rollup.config.js -f es -o dist/ranger.min.esm.js",
-    "build:iife": "../node_modules/.bin/rollup src/main.js -c ../rollup.config.js -f iife -o dist/ranger.iife.js",
-    "build:iife:min": "../node_modules/.bin/rollup src/main.js -c ../rollup.config.js -f iife -o dist/ranger.min.iife.js",
     "test": "npm run --prefix .. test",
     "prepublishOnly": "npm run build && npm run test"
   },
   "bundlesize": [
     {
-      "path": "dist/ranger.min.iife.js",
+      "path": "dist/ranger.min.esm.js",
       "maxSize": "5.5 kB"
     }
   ],
