+++ ranger/src/RangePicker.html
@@ -33,9 +33,9 @@
 <script>
   import { calendarMonth, getNextMonth } from 'cntdys'
   import { dayClass, selectDay } from './functions.js'
-  import { weekClass } from '../../common/src/index.js'
-  import Wrapper from '../../common/src/Wrapper.html'
-  import Header from '../../common/src/Header.html'
+  import { weekClass } from '../../common/index.js'
+  import Wrapper from '../../common/Wrapper.html'
+  import Header from '../../common/Header.html'
 
   export default {
     helpers: {
