+++ rollup.config.js
@@ -6,7 +6,7 @@ const production = !process.env.ROLLUP_WATCH
 const outputIndex = production ? 8 : 9 // TODO: make this more robust
 const ranger = process.argv[outputIndex].includes('ranger')
 const min = process.argv[outputIndex].includes('.min')
-const name = ranger ? 'Ranger' : 'Nanocal' // will be used in iife
+const name = ranger ? 'Ranger' : 'Nanocal'
 const cssName = ranger ? 'ranger' : 'nanocal'
 
 export default {
@@ -18,7 +18,7 @@ export default {
     resolve(),
     svelte({
       dev: !production, // enable run-time checks when not in production
-      css: css => { css.write(`dist/${cssName}.min.css`) }
+      css: css => { css.write(`build/${cssName}.min.css`) }
     }),
     production && min && minify({ comments: false })
   ]
