+++ test/ranger/functions.test.ts
@@ -1,7 +1,7 @@
 import * as test from 'tape'
 import { dayClass, selectDay } from '../../ranger/src/functions'
 import { IDay } from 'cntdys'
-import { ICalendarDay } from '../../common/src'
+import { ICalendarDay } from '../../common'
 
 test('dayClass', assert => {
   let weekDay: IDay = {
